chore(gateway): address tool activity review feedback
