+++ crates/ironclaw_gateway/static/js/core/tool-activity.test.mjs
@@ -3,123 +3,15 @@ import { readFileSync } from "node:fs";
 import test from "node:test";
 import vm from "node:vm";
 
-class FakeClassList {
-  constructor(owner) {
-    this.owner = owner;
-  }
+import {
+  chatChildKinds,
+  createChatContainer,
+  createMessageElement,
+  FakeElement,
+} from "./test-dom.mjs";
 
-  values() {
-    return String(this.owner.className || "")
-      .split(/\s+/)
-      .filter(Boolean);
-  }
-
-  contains(name) {
-    return this.values().includes(name);
-  }
-
-  add(...names) {
-    const next = new Set(this.values());
-    for (const name of names) next.add(name);
-    this.owner.className = Array.from(next).join(" ");
-  }
-
-  remove(...names) {
-    const remove = new Set(names);
-    this.owner.className = this.values()
-      .filter((name) => !remove.has(name))
-      .join(" ");
-  }
-
-  toggle(name, force) {
-    const has = this.contains(name);
-    const shouldAdd = force === undefined ? !has : Boolean(force);
-    if (shouldAdd) this.add(name);
-    else this.remove(name);
-    return shouldAdd;
-  }
-}
-
-class FakeElement {
-  constructor(tagName) {
-    this.tagName = tagName;
-    this.children = [];
-    this.parentNode = null;
-    this.className = "";
-    this.classList = new FakeClassList(this);
-    this.attributes = new Map();
-    this.style = {};
-    this.textContent = "";
-    this._innerHTML = "";
-  }
-
-  appendChild(child) {
-    if (child.parentNode) child.parentNode.removeChild(child);
-    child.parentNode = this;
-    this.children.push(child);
-    return child;
-  }
-
-  insertBefore(child, anchor) {
-    if (!anchor) return this.appendChild(child);
-    const index = this.children.indexOf(anchor);
-    if (index < 0) return this.appendChild(child);
-    if (child.parentNode) child.parentNode.removeChild(child);
-    child.parentNode = this;
-    this.children.splice(index, 0, child);
-    return child;
-  }
-
-  removeChild(child) {
-    const index = this.children.indexOf(child);
-    if (index >= 0) this.children.splice(index, 1);
-    child.parentNode = null;
-    return child;
-  }
-
-  remove() {
-    if (this.parentNode) this.parentNode.removeChild(this);
-  }
-
-  setAttribute(name, value) {
-    this.attributes.set(name, String(value));
-  }
-
-  getAttribute(name) {
-    return this.attributes.get(name) || null;
-  }
-
-  removeAttribute(name) {
-    this.attributes.delete(name);
-  }
-
-  addEventListener() {}
-
-  querySelector() {
-    return null;
-  }
-
-  set innerHTML(value) {
-    this._innerHTML = String(value);
-    this.children = [];
-  }
-
-  get innerHTML() {
-    return this._innerHTML;
-  }
-}
-
-function message(role) {
-  const element = new FakeElement("div");
-  element.className = `message ${role}`;
-  return element;
-}
-
-function createHarness(initialChildren) {
-  const container = new FakeElement("div");
-  container.className = "chat-messages";
-  for (const child of initialChildren) container.appendChild(child);
-
+function createToolActivityHarness(initialChildren = []) {
+  const container = createChatContainer(initialChildren);
   const context = {
     clearInterval: () => {},
     Date,
@@ -145,23 +37,10 @@ function createHarness(initialChildren) {
   };
 }
 
-function childKinds(container) {
-  return container.children.map((child) => {
-    if (child.classList.contains("activity-group")) return "activity";
-    if (child.classList.contains("message")) {
-      if (child.classList.contains("assistant")) return "assistant";
-      if (child.classList.contains("user")) return "user";
-      if (child.classList.contains("system")) return "system";
-      return "message";
-    }
-    return child.tagName;
-  });
-}
-
 test("tool activity started after a trailing assistant reply renders before that reply", () => {
-  const { container, controller } = createHarness([
-    message("user"),
-    message("assistant"),
+  const { container, controller } = createToolActivityHarness([
+    createMessageElement("user"),
+    createMessageElement("assistant"),
   ]);
 
   controller.startTool({
@@ -169,7 +48,7 @@ test("tool activity started after a trailing assistant reply renders before that
     name: "extension_search",
   });
 
-  assert.deepEqual(childKinds(container), ["user", "activity", "assistant"]);
+  assert.deepEqual(chatChildKinds(container), ["user", "activity", "assistant"]);
 
   controller.completeTool({
     call_id: "call-extension-search",
@@ -179,14 +58,14 @@ test("tool activity started after a trailing assistant reply renders before that
   });
   controller.finalizeGroup();
 
-  assert.deepEqual(childKinds(container), ["user", "activity", "assistant"]);
+  assert.deepEqual(chatChildKinds(container), ["user", "activity", "assistant"]);
 });
 
 test("tool activity after a follow-up user message stays with the active follow-up turn", () => {
-  const { container, controller } = createHarness([
-    message("user"),
-    message("assistant"),
-    message("user"),
+  const { container, controller } = createToolActivityHarness([
+    createMessageElement("user"),
+    createMessageElement("assistant"),
+    createMessageElement("user"),
   ]);
 
   controller.startTool({
@@ -194,7 +73,7 @@ test("tool activity after a follow-up user message stays with the active follow-
     name: "calendar",
   });
 
-  assert.deepEqual(childKinds(container), [
+  assert.deepEqual(chatChildKinds(container), [
     "user",
     "assistant",
     "user",
@@ -205,9 +84,9 @@ test("tool activity after a follow-up user message stays with the active follow-
 test("tool activity does not skip non-message cards after an assistant reply", () => {
   const card = new FakeElement("div");
   card.className = "auth-card";
-  const { container, controller } = createHarness([
-    message("user"),
-    message("assistant"),
+  const { container, controller } = createToolActivityHarness([
+    createMessageElement("user"),
+    createMessageElement("assistant"),
     card,
   ]);
 
@@ -216,10 +95,21 @@ test("tool activity does not skip non-message cards after an assistant reply", (
     name: "drive",
   });
 
-  assert.deepEqual(childKinds(container), [
+  assert.deepEqual(chatChildKinds(container), [
     "user",
     "assistant",
     "div",
     "activity",
   ]);
+});
+
+test("tool activity appends normally when there is no existing chat message", () => {
+  const { container, controller } = createToolActivityHarness();
+
+  controller.startTool({
+    call_id: "call-empty",
+    name: "bootstrap",
+  });
+
+  assert.deepEqual(chatChildKinds(container), ["activity"]);
 });
